\name{DOBSUN}
\alias{DOBSUN}
\alias{DOBSUN}
\docType{package}
\title{
Utilities for DOBSUN, a combined Dobsonmeter-sun photometer
}
\description{
Utilities for calculating column ozone concentration from Dobson meter data and column aerosol optical thickness from sun photometer data.
}
\details{
\tabular{ll}{
Package: \tab DOBSUN\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-02-14\cr
License: \tab GPL-3\cr
}
}

\author{
Al Fischer
Maintainer: Al Fischer <alphonse.fisch@gmail.com>
}
\references{
  http://www.wmo.int/pages/prog/arep/gaw/documents/GAW183-Dobson-WEB.pdf
}

\keyword{ package }

